CVS-173 スキーマ定義の変更に対応

+++ pkg/qemu/type.go
@@ -3,7 +3,7 @@ package qemu
 type InstallOpts struct {
 	Name   string
 	Memory uint64
-	VCpu   uint64
+	VCpu   int
 	Image  string
 	Disk   string
 }
